+++ lib/rdf/rdfa.rb
@@ -29,7 +29,11 @@ module RDF
     require 'rdf/rdfa/patches/nokogiri_hacks'
     autoload :Profile,  'rdf/rdfa/profile'
     autoload :Reader,  'rdf/rdfa/reader'
+    autoload :Writer,  'rdf/rdfa/writer'
     autoload :VERSION, 'rdf/rdfa/version'
+    
+    XML_RDFA_PROFILE    = "http://www.w3.org/profile/rdfa-1.1"
+    XHTML_RDFA_PROFILE  = "http://www.w3.org/profile/html-rdfa-1.1"
     
     def self.debug?; @debug; end
     def self.debug=(value); @debug = value; end