+++ spec/matchers.rb
@@ -1,3 +1,31 @@
+require 'rspec/matchers'
+
+RSpec::Matchers.define :have_xpath do |xpath, value|
+  match do |actual|
+    @doc = Nokogiri::XML.parse(actual)
+    @doc.should be_a(Nokogiri::XML::Document)
+    @doc.root.should be_a(Nokogiri::XML::Element)
+    @namespaces = @doc.namespaces.merge("xhtml" => "http://www.w3.org/1999/xhtml", "xml" => "http://www.w3.org/XML/1998/namespace")
+    case value
+    when false
+      @doc.root.at_xpath(xpath, @namespaces).should be_nil
+    when true
+      @doc.root.at_xpath(xpath, @namespaces).should_not be_nil
+    when Array
+      @doc.root.at_xpath(xpath, @namespaces).to_s.split(" ").should include(*value)
+    when Regexp
+      @doc.root.at_xpath(xpath, @namespaces).to_s.should =~ value
+    else
+      @doc.root.at_xpath(xpath, @namespaces).to_s.should == value
+    end
+  end
+  
+  failure_message_for_should do |actual|
+    msg = "expected to that #{xpath.inspect} would be #{value.inspect} in:\n" + actual.to_s
+    msg += "was: #{@doc.root.at_xpath(xpath, @namespaces)}"
+  end
+end
+
 module Matchers
   class BeEquivalentGraph
     Info = Struct.new(:about, :information, :trace, :compare, :inputDocument, :outputDocument)
@@ -31,7 +59,7 @@ module Matchers
 
     def matches?(actual)
       @actual = normalize(actual)
-      @actual == @expected
+      @actual.isomorphic_with?(@expected)
     end
 
     def failure_message_for_should
@@ -46,8 +74,8 @@ module Matchers
       "\n#{info + "\n" unless info.empty?}" +
       (@info.inputDocument ? "Input file: #{@info.inputDocument}\n" : "") +
       (@info.outputDocument ? "Output file: #{@info.outputDocument}\n" : "") +
-      "Unsorted Expected:\n#{@expected.to_ntriples}" +
-      "Unsorted Results:\n#{@actual.to_ntriples}" +
+      "Unsorted Expected:\n#{@expected.dump(:ntriples)}" +
+      "Unsorted Results:\n#{@actual.dump(:ntriples)}" +
       (@info.trace ? "\nDebug:\n#{@info.trace}" : "")
     end
     def negative_failure_message
@@ -58,7 +86,6 @@ module Matchers
   def be_equivalent_graph(expected, info = nil)
     BeEquivalentGraph.new(expected, info)
   end
-
   # Run expected SPARQL query against actual
   if $redland_enabled
     class PassQuery
@@ -73,7 +100,7 @@ module Matchers
         @expected_results = @info.respond_to?(:expectedResults) ? @info.expectedResults : true
         model = Redland::Model.new
         ntriples_parser = Redland::Parser.ntriples
-        ntriples_parser.parse_string_into_model(model, actual.to_ntriples, "http://www.w3.org/2006/07/SWD/RDFa/testsuite/xhtml1-testcases/")
+        ntriples_parser.parse_string_into_model(model, actual.dump(:ntriples), "http://www.w3.org/2006/07/SWD/RDFa/testsuite/xhtml1-testcases/")
 
         @results = @query.execute(model)
         #puts "Redland query results: #{@results.inspect}"
@@ -97,7 +124,7 @@ module Matchers
         end +
         "\n#{@expected}" +
         "\n#{@info.input}" +
-        "\nResults:\n#{@actual.to_ntriples}" +
+        "\nResults:\n#{@actual.dump(:ntriples)}" +
         "\nDebug:\n#{@info.trace}"
       end
     end
@@ -107,65 +134,5 @@ module Matchers
     end
   else
     def pass_query(expect, info = ""); false; end
-  end
-
-  class BeValidXML
-    def initialize(info)
-      @info = info
-    end
-    def matches?(actual)
-      @actual = actual
-      @doc = Nokogiri::XML.parse(actual)
-      @results = @doc.validate
-      @results.nil?
-    rescue
-      false
-    end
-    def failure_message_for_should
-      "#{@info + "\n" unless @info.empty?}" +
-      if @doc.nil?
-        "did not parse"
-      else
-        "\n#{@results}" +
-        "\nParsed:\n#{@doc}"
-      end   +
-        "\nActual:\n#{@actual}"
-    end
-  end
-  
-  def be_valid_xml(info = "")
-    BeValidXML.new(info)
-  end
-
-  class BeEquivalentXML
-    def initialize(expected, info)
-      @expected = expected
-      @info = info
-    end
-    
-    def matches?(actual)
-      @actual = actual
-
-      a = @actual.index("<") == 0 ? @actual : "<foo>#{@actual}</foo>"
-      e = @expected.index("<") == 0 ? @expected : "<foo>#{@expected}</foo>"
-      a_hash = ActiveSupport::XmlMini.parse(a)
-      e_hash = ActiveSupport::XmlMini.parse(e)
-      a_hash == e_hash
-    rescue
-      puts $!
-      @fault = $!.message
-      false
-    end
-
-    def failure_message_for_should
-      "#{@info + "\n" unless @info.empty?}" +
-      "Fault: #{@fault + "\n" if @fault}" +
-      "Expected:#{@expected}\n" +
-      "Actual:#{@actual}"
-    end
-  end
-  
-  def be_equivalent_xml(expected, info = "")
-    BeEquivalentXML.new(expected, info)
   end
 end